% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{bootstrap}
\alias{bootstrap}
\title{Generate bootstrap replicates}
\usage{
bootstrap(n, times = 1L, weights = NULL, bayes = FALSE, size = n)
}
\arguments{
\item{n}{A positive, scalar integer representing the number
of observations (items to choose from).}

\item{times}{A positive, scalar integer representing the number of
bootstrap samples to draw.}

\item{weights}{A numeric vector with observation level weights.}

\item{bayes}{A scalar logical, indicating whether to use the Bayesian
bootstrap.}

\item{size}{The number of observations in each output sample.}
}
\description{
Generate nonparametric bootstrap samples. In addition to the the ordinary
bootstrap, it supports weights, the Bayesian bootstrap, and sub- or
super-sampling.
}
\examples{
# ordinary non-parametric bootstrap
bootstrap(20, 5)

# Bayesian bootstrap
bootstrap(20, 5, bayes = TRUE)

# weighted bootstrap
w <- c(rep(5, 5), rep(1, 15))
bootstrap(20, 5, weights = w / sum(w))
}
\references{
\itemize{
\item Angelo Canty and Brian Ripley (2016). "boot: Bootstrap R (S-Plus) Functions." R package version 1.3-18.
\item Davison, A. C. & Hinkley, D. V. (1997) Bootstrap Methods and Their Applications. Cambridge University Press, Cambridge. ISBN 0-521-57391-2
}
}
\seealso{
The \pkg{boot} function \code{\link[boot:boot]{boot::boot()}} which is the
canonical R bootstrap implementation. The \pkg{modelr} function
\code{\link[modelr:bootstrap]{modelr::bootstrap()}}, \pkg{rsample} function \link[rsample:bootstraps]{rsample::bootstraps}
provide similar implementations specialized to data frames.
}
