% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_balanced.R
\name{bootstrap_balanced}
\alias{bootstrap_balanced}
\title{Generate balanced bootstrap replicates}
\usage{
bootstrap_balanced(n, times = 1L)
}
\arguments{
\item{n}{A positive, scalar integer representing the number
of observations (items to choose from).}

\item{times}{A positive, scalar integer representing the number of
bootstrap samples to draw.}
}
\description{
A balanced bootstrap generates \code{times} replicates in which each element
appears \code{times} times.
}
\examples{
bootstrap_balanced(20, 5)
}
\references{
\itemize{
\item Gleason, John 1988. "Algorithms for Balanced Bootstrap Simulations".
\emph{The American Statistician}.
\href{https://dx.doi.org/10.2307/2685134}{doi:10.2307/2685134}.
\item Davison, A.C., Hinkley, D. V., and Schechtman, E. 1986.
"Efficient Bootstrap Simulation." \emph{Biometrika}.
\item Angelo Canty and Brian Ripley (2016).
\emph{boot: Bootstrap R (S-Plus) Functions}. R package version 1.3-18.
}
}
\seealso{
The \pkg{boot} function \code{\link[boot:boot]{boot::boot()}} which is the
canonical R bootstrap implementation.
}
