% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_pois.R
\name{bootstrap_pois}
\alias{bootstrap_pois}
\title{Generate Poisson bootstrap replicates}
\usage{
bootstrap_pois(n, times)
}
\arguments{
\item{n}{A positive, scalar integer representing the number
of observations (items to choose from).}

\item{times}{A positive, scalar integer representing the number of
bootstrap samples to draw.}
}
\description{
Generate Poisson bootstrap replicates
}
