% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_ts.R
\name{bootstrap_ts}
\alias{bootstrap_ts}
\title{Generate time-series bootstrap replicates}
\usage{
bootstrap_ts(n, times = 1L, block_size = 1L, size = n, type = c("fixed",
  "geom"), endcorr = TRUE)
}
\arguments{
\item{n}{The number of observations.}

\item{times}{The number of bootstrap samples.}

\item{block_size}{If \code{type = "fixed"}, then \code{size} is the fixed block length
used in generating the bootstrap samples.  If \code{type = "geom"}
then \code{size} is the mean of the geometric distribution from which the
block lengths were sample. The value of \code{size} must be a positive
integer less than \code{n}.}

\item{size}{A scalar integer representing the number of items in the
output samples.}

\item{type}{The type of simulation used to generate the replicate time
series. The possible input values are \code{"fixed"}, for block resampling
with fixed block lengths of \code{size}, and \code{"geom"}, for block resampling
with block lengths having a geometric distribution with mean \code{size}).}

\item{endcorr}{A scalar logical indicating whether to adjust for end
correlations in blocks.}
}
\description{
Generate bootstrap replicates for time series items.
The replicate time series can be generated using either fixed or random
block lengths.
}
\details{
If \code{type = "fixed"} then each replicate time series is found
by taking blocks of length \code{block_size}, from the original time series and
putting them end-to-end until a new series of length \code{size} is created.
When \code{type = "geom"}, a similar approach is taken except
that the block lengths are sampled from a geometric distribution with mean
\code{block_size}.
}
\references{
\itemize{
\item Kunsch, H.R. (1989) The jackknife and the bootstrap for general stationary observations. \emph{Annals of Statistics}
\item Davison, A.C. and Hinkley, D.V. (1997) \emph{Bootstrap Methods and Their Application}. Cambridge University Press.
\item Politis, D.N. and Romano, J.P. (1994) "The stationary bootstrap. Journal of the American Statistical Association."
\item Angelo Canty and Brian Ripley (2016). boot: Bootstrap R (S-Plus) Functions. R package version 1.3-18.
}
}
\seealso{
The \pkg{boot} function \code{\link[boot:tsboot]{boot::tsboot()}}, from which this function is
derived.
}
