% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossv_vfold_ts.R
\name{cross_vfold_ts}
\alias{cross_vfold_ts}
\title{Generate cross-validated time-series K-fold test-training pairs}
\usage{
cross_vfold_ts(n, v = 5L)
}
\arguments{
\item{n}{A positive, scalar integer representing the number
of observations (items to choose from).}

\item{v}{A scalar integer representing the number of folds to generate.}
}
\description{
Splits data into test/train sets for cross-validating time series
data. In each split, the indices are increasing, so only previous data
are used to test future data.
}
\details{
The data are split into \eqn{K \geq 2} folds, where for each
\eqn{i \in 2:K}{i in 2:K}, the \eqn{i}-th fold is the test set
and folds \eqn{1, ..., (i - 1)} folds are the training set.
}
\references{
\itemize{
\item Scikit-learn v. 0.18.1. Cross-validation: evaluating estimator performance.
\href{http://scikit-learn.org/stable/modules/cross_validation.html#cross-validation}{URL}.
}
}
