% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossv_lpo.R
\name{crossv_lpo}
\alias{crossv_lpo}
\alias{crossv_loo}
\title{Generate cross-validated leave-p-out test/training pairs}
\usage{
crossv_lpo(n, size = 1L)

crossv_loo(n)
}
\arguments{
\item{n}{A positive, scalar integer representing the number
of observations (items to choose from).}

\item{size}{A scalar integer representing the number of items to include
in the training set.}
}
\description{
Generate cross-validated leave-one-out or leave-p-out test/training pairs.
The function \code{leave-p-out} generates leave-p-out test/training pairs.
The function \code{leave-one-out} is convenience function for the
common special case of leave-one-out cross-validation, \code{p = 1}.
}
\examples{
# Example originally from modelr::crossv_mc
library("purrr")
# leave-on-out cross-validation
crossv_loo(10)

# leave-two-out cross-validation
crossv_lpo(10, size = 2)
}
