% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossv_ts.R
\name{crossv_ts}
\alias{crossv_ts}
\title{Generate test/train sets for time-series}
\usage{
crossv_ts(n, horizon = 1L, test_size = 1L, test_partial = FALSE,
  train_partial = TRUE, train_size = n, test_start = NULL, from = 1L,
  to = n, by = 1L)
}
\arguments{
\item{n}{A positive, scalar integer representing the number
of observations (items to choose from).}

\item{horizon}{Difference between the first test set observation and the last training set observation}

\item{test_size}{A scalar integer indicating the size of the test set.}

\item{test_partial, train_partial}{A logical scalar or positive scalar integer which
handles how partial test (train) sets are handled. If \code{TRUE}, then partial
test (train) sets are allowed. If \code{FALSE}, then partial test (train) sets
are not allowed. If it is an integer, then it is the minimum allowable
size of a test (train) set.}

\item{train_size}{The maximum size of the training set. This allows for a
rolling window training set, instead of using all observations from the
start of the time series.}

\item{test_start}{An integer vector of the starting index values of the
test set.}

\item{from, to, by}{Integer scalar values used to generate values of
\code{test_start} if \code{test_start = NULL}. In that case,
\code{test_start = seq(from, to, by)}.}
}
\description{
Generate test/train sets for time series. Each training set consists of
observations before the test set. This method also called "evaluation on a
rolling forecasting origin".
}
\details{
In time-series cross-validation the training set only uses
observations that are prior to the test set. Suppose the time series has
\eqn{n} observations, the training set has a maximum size of
\eqn{r \leq n}{r <= n} and minimum size of \eqn{s \geq r}{s >= r}.
and the test set has a maximum size of \eqn{p \leq n}{p <= n} and
minimum size of \eqn{q \geq p}{q >= p}.

For indices \eqn{i \in \{1, \dots, N\}}{i in \{1, \dots, N\}}:
\itemize{
\item Select observations \eqn{i, \dots, \max(p, n)}{i, ..., max(p, n)}
for the test set.
\item Select observations
\eqn{\max(i - h - p), \dots, i - h}{max(i - h - p), ..., i - h}
for the training set.
\item If the test set has a size of at least \code{q} and the training set
has a size of at least \code{r}.
}
}
\references{
\itemize{
\item Hyndman RJ (2017).
\emph{forecast: Forecasting functions for time series and linear models}.
R package version 8.0,
\href{http://github.com/robjhyndman/forecast}{URL}.
\item Hyndman RJ and Khandakar Y (2008). "Automatic time series forecasting:
the forecast package for R." \emph{Journal of Statistical Software}.
\href{http://www.jstatsoft.org/article/view/v027i03}{URL}.
\item Rob J. Hyndman.
"\href{http://robjhyndman.com/hyndsight/tscv/}{Cross-validation for time series}".
December 5, 2016.
\item Rob J. Hyndman.
"\href{http://robjhyndman.com/hyndsight/tscvexample/}{Time series cross-validation: an R example}".
August 26, 2011.
\item Rob J. Hyndman and George Athanasopoulos.
"Evaluating Forecast Accuracy." \href{https://www.otexts.org/fpp/2/5}{URL}.
\item Max Kuhn. "Data splitting for Time Series." \emph{The caret Package}.
2016-11-29. \href{https://topepo.github.io/caret/data-splitting.html}{URL}.
}
}
