% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossv_vfold.R
\name{crossv_vfold}
\alias{crossv_vfold}
\title{Generate cross-validated v-fold test-training pairs}
\usage{
crossv_vfold(n, v = 5L, shuffle = TRUE)
}
\arguments{
\item{n}{A positive, scalar integer representing the number
of observations (items to choose from).}

\item{v}{A scalar integer representing the number of folds in cross
validation.}

\item{shuffle}{A flag which indicates whether to randomly assign
observations to folds.}
}
\description{
Generate cross-validated v-fold test-training pairs
}
\examples{
crossv_vfold(20, v = 5)
}
\references{
\itemize{
\item Breiman, L., Friedman, J.H., Olshen, R.A. and Stone, C.J. (1984) \emph{Classification and Regression Trees}. Wadsworth.
\item Burman, P. (1989) "A comparative study of ordinary cross-validation,
v-fold cross-validation and repeated learning-testing methods."
\emph{Biometrika}, 76, 503–514
\item Davison, A.C. and Hinkley, D.V. (1997)
\emph{Bootstrap Methods and Their Application}. Cambridge University Press.
\item Efron, B. (1986) "How biased is the apparent error rate of a
prediction rule?" \emph{Journal of the American Statistical Association},
81, 461–470.
\item Stone, M. (1974) "Cross-validation choice and assessment of statistical
predictions (with Discussion)".
\emph{Journal of the Royal Statistical Society}, B, 36, 111–147.
}
}
