% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jackknife.R
\name{jackknife}
\alias{jackknife}
\title{Generate Jackknife Subsamples}
\usage{
jackknife(n, size = 1L)
}
\arguments{
\item{n}{A positive, scalar integer representing the number
of observations (items to choose from).}

\item{size}{A scalar integer representing the number of elements to delete.
When \code{size = 1}, the subsets are generated by ordinary
(delete-one) jackknife. When \code{size > 1}, the subsets are generated by
the block (delete-\eqn{p}) jackknife.}
}
\description{
Generate jackknife subsamples. For data with \eqn{n} elements,
the jackknife generates \eqn{n} subsamples, each with one element
deleted. The generalized (delete-\eqn{p}, block) jackknife generates
subsamples by deleting \code{p} observations.
}
\references{
\itemize{
\item Davison, A. C. & Hinkley, D. V. (1997)
\emph{Bootstrap Methods and Their Applications}.
Cambridge University Press, Cambridge. ISBN 0-521-57391-2.
\item Tukey, J. W. (1958). "Bias and confidence in not quite large samples".
\emph{The Annals of Mathematical Statistics}.
\href{https://dx.doi.org/10.1214/aoms/1177706647}{doi:10.1214/aoms/1177706647}.
\item Efron, B.; Stein, C. (May 1981). "The Jackknife Estimate of Variance".
\emph{The Annals of Statistics}.
\href{https://dx.doi.org/10.1214/aos/1176345462}{doi:10.1214/aos/1176345462}.
}
}
