% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permute.R
\name{permute}
\alias{permute}
\title{Generate permutation subsets}
\usage{
permute(n, times = 1L, force_random = FALSE)
}
\arguments{
\item{n}{A positive, scalar integer representing the number
of observations (items to choose from).}

\item{times}{A positive, scalar integer representing the number of
bootstrap samples to draw.}

\item{force_random}{If \code{TRUE}, then generate random permutations, even
if \code{n} is small enough that \code{R} is more than the number of
permutations.}
}
\description{
Generate \code{R} replicates with permutations of the rows or groups of a data frame.
}
\details{
If \code{R} is greater than the total number of permutations,
only the total number of permutations is returned. Since the total number
of permutations is \code{factorial(n)}, this will only happen if \code{n}
is small.
}
