% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roll.R
\name{roll}
\alias{roll}
\title{Generate rolling windows}
\usage{
roll(n, width = 1L, align = c("left", "right", "center"), partial = TRUE,
  indices = NULL, from = 1L, to = n, by = 1L, offsets = NULL)
}
\arguments{
\item{n}{A positive, scalar integer representing the number
of observations (items to choose from).}

\item{width}{The window width.}

\item{align}{Is the window left-aligned, right-aligned, or centered
relative to the reference index of the window.}

\item{partial}{A logical flag or numeric vector. If \code{FALSE} all indexes of a rolling
window must be within the rows (or groups) of \code{data}. If \code{TRUE},
then the subset of indexes that are in range are used.
A numeric argument to \code{partial} can be used to determine the minimal
window size for partial windows.}

\item{indices, from, to, by}{The indices at which to generate windows. If
\code{indices} is non-\code{NULL}, then it is used. Otherwise, the indices
are generated from \code{seq(from, to, by)}.}

\item{offsets}{An integer vector or \code{NULL}. If non-\code{NULL},
then it is the offsets of elements in the window relative to the
reference index of the window; \code{0} is the reference index, negative
numbers are indices before the index, while positive numbers are after the
index. If \code{NULL}, the offsets are generated from the \code{width}
and \code{align} arguments.}
}
\description{
Generate resample objects of rolling windows over elements of a data frame.
}
